organize stats components and improve data loading
