+++ components/stats/StatsPage.tsx
@@ -6,18 +6,19 @@ import { useViewport } from 'hooks/useViewport'
 import { useRouter } from 'next/router'
 import { useEffect, useMemo, useState } from 'react'
 import { breakpoints } from 'utils/theme'
-import MangoStats from './MangoStats'
-import PerpStats from './PerpStats'
-import PerpStatsPage from './PerpStatsPage'
-import SpotMarketsTable from './SpotMarketsTable'
-import TokenStats from './TokenStats'
+import MangoStats from './mango/MangoStats'
+import PerpStats from './perps/PerpStats'
+import PerpStatsPage from './perps/PerpStatsPage'
+import SpotMarketsTable from './spot/SpotMarketsTable'
+import TokenStats from './tokens/TokenStats'
 
 const TABS = ['tokens', 'perp-markets', 'spot-markets', 'mango-stats']
+const actions = mangoStore.getState().actions
 
 const StatsPage = () => {
   const [activeTab, setActiveTab] = useState('tokens')
-  const actions = mangoStore.getState().actions
   const perpStats = mangoStore((s) => s.perpStats.data)
+  const initialStatsLoad = mangoStore((s) => s.tokenStats.initialLoad)
   const perpPositionsStatsNotLoaded = mangoStore(
     (s) => s.perpStats.positions.initialLoad,
   )
@@ -39,6 +40,13 @@ const StatsPage = () => {
       actions.fetchPositionsStats()
     }
   }, [group, perpPositionsStatsNotLoaded])
+
+  useEffect(() => {
+    if (group && !initialStatsLoad) {
+      const actions = mangoStore.getState().actions
+      actions.fetchTokenStats()
+    }
+  }, [group, initialStatsLoad])
 
   const tabsWithCount: [string, number][] = useMemo(() => {
     return TABS.map((t) => [t, 0])
