+++ components/stats/perps/AverageFundingChart.tsx
