+++ components/stats/perps/PerpMarketDetails.tsx
