+++ components/stats/spot/SpotMarketsTable.tsx
@@ -1,9 +1,9 @@
 import { useTranslation } from 'next-i18next'
 import { useCallback } from 'react'
-import { useViewport } from '../../hooks/useViewport'
-import { COLORS } from '../../styles/colors'
-import { breakpoints } from '../../utils/theme'
-import ContentBox from '../shared/ContentBox'
+import { useViewport } from '../../../hooks/useViewport'
+import { COLORS } from '../../../styles/colors'
+import { breakpoints } from '../../../utils/theme'
+import ContentBox from '../../shared/ContentBox'
 import MarketLogos from '@components/trade/MarketLogos'
 import useMangoGroup from 'hooks/useMangoGroup'
 import {
