+++ components/stats/tokens/TokenOverviewTable.tsx
@@ -1,10 +1,10 @@
 import { Disclosure, Transition } from '@headlessui/react'
 import { ChevronDownIcon, ChevronRightIcon } from '@heroicons/react/20/solid'
 import { useTranslation } from 'next-i18next'
-import { useViewport } from '../../hooks/useViewport'
-import { breakpoints } from '../../utils/theme'
-import { LinkButton } from '../shared/Button'
-import ContentBox from '../shared/ContentBox'
+import { useViewport } from '../../../hooks/useViewport'
+import { breakpoints } from '../../../utils/theme'
+import { LinkButton } from '../../shared/Button'
+import ContentBox from '../../shared/ContentBox'
 import Tooltip from '@components/shared/Tooltip'
 import { Bank, toUiDecimals } from '@blockworks-foundation/mango-v4'
 import { NextRouter, useRouter } from 'next/router'
